Move ProteinServiceImpl and AnnotationView to Protein. Move GO reading methods back to MS2.
