+++ ms2/src/org/labkey/ms2/peptideview/AbstractMS2RunView.java
@@ -39,8 +39,8 @@ import org.labkey.api.data.TSVGridWriter;
 import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.TableSelector;
-import org.labkey.api.protein.ProteinDictionaryHelpers;
 import org.labkey.api.protein.ProteinManager;
+import org.labkey.api.protein.go.GoLoader;
 import org.labkey.api.query.DetailsURL;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QueryNestingOption;
@@ -67,7 +67,7 @@ import org.labkey.ms2.PeptideManager;
 import org.labkey.ms2.RunListException;
 import org.labkey.ms2.SpectrumIterator;
 import org.labkey.ms2.SpectrumRenderer;
-import org.labkey.api.protein.go.GoLoader;
+import org.labkey.ms2.protein.tools.GoHelpers;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -164,11 +164,11 @@ public abstract class AbstractMS2RunView
         if (GoLoader.isGoLoaded())
         {
             MenuButton goButton = new MenuButton("Gene Ontology Charts");
-            List<ProteinDictionaryHelpers.GoTypes> types = new ArrayList<>();
-            types.add(ProteinDictionaryHelpers.GoTypes.CELL_LOCATION);
-            types.add(ProteinDictionaryHelpers.GoTypes.FUNCTION);
-            types.add(ProteinDictionaryHelpers.GoTypes.PROCESS);
-            for (ProteinDictionaryHelpers.GoTypes goType : types)
+            List<GoHelpers.GoTypes> types = new ArrayList<>();
+            types.add(GoHelpers.GoTypes.CELL_LOCATION);
+            types.add(GoHelpers.GoTypes.FUNCTION);
+            types.add(GoHelpers.GoTypes.PROCESS);
+            for (GoHelpers.GoTypes goType : types)
             {
                 ActionURL url = MS2Controller.getPeptideChartURL(getContainer(), goType);
                 goButton.addMenuItem(goType.toString(), dataRegion.getJavascriptFormReference() + ".action=\"" + url.getLocalURIString() + "\"; " + dataRegion.getJavascriptFormReference() + ".submit();");
