+++ ms2/src/org/labkey/ms2/protein/tools/GOPieToolTipGenerator.java
@@ -18,7 +18,6 @@ package org.labkey.ms2.protein.tools;
 
 import org.jfree.chart.labels.PieToolTipGenerator;
 import org.jfree.data.general.PieDataset;
-import org.labkey.api.protein.ProteinDictionaryHelpers;
 
 /**
  * User: tholzman
@@ -39,7 +38,7 @@ public class GOPieToolTipGenerator implements PieToolTipGenerator
         String tip = null;
         try
         {
-            tip = ProteinDictionaryHelpers.getGODefinitionFromAcc(acc);
+            tip = GoHelpers.getGODefinitionFromAcc(acc);
         }
         catch (Exception e)
         {
